+++ agents/agents/agentsintegration/agentsintegration_test.go
@@ -8,11 +8,14 @@ import (
 
 	awsTime "github.com/aws/smithy-go/time"
 	"github.com/brianvoe/gofakeit/v6"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
+	. "github.com/stretchr/testify/assert"
 	"github.com/synapsecns/sanguine/agents/agents/executor"
-	executorCfg "github.com/synapsecns/sanguine/agents/agents/executor/config"
 	"github.com/synapsecns/sanguine/agents/agents/guard"
 	"github.com/synapsecns/sanguine/agents/agents/notary"
+	"github.com/synapsecns/sanguine/agents/config"
+	execConfig "github.com/synapsecns/sanguine/agents/config/executor"
 	"github.com/synapsecns/sanguine/agents/types"
 	config2 "github.com/synapsecns/sanguine/ethergo/signer/config"
 
@@ -35,8 +38,6 @@ func RemoveAgentsTempFile(t *testing.T, fileName string) {
 
 //nolint:cyclop,maintidx
 func (u *AgentsIntegrationSuite) TestAgentsE2E() {
-	// TODO (joe and lex): FIX ME
-	// u.T().Skip()
 	testDone := false
 	defer func() {
 		testDone = true
@@ -49,7 +50,7 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 	summitClient, err := backfill.DialBackend(u.GetTestContext(), u.TestBackendSummit.RPCAddress(), u.ScribeMetrics)
 	u.Nil(err)
 
-	originConfig := scribeConfig2.ContractConfig{
+	originConfig := scribeConfig.ContractConfig{
 		Address:    u.OriginContract.Address().String(),
 		StartBlock: 0,
 	}
@@ -65,7 +66,7 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 		},
 		Contracts: []scribeConfig2.ContractConfig{originConfig},
 	}
-	destinationConfig := scribeConfig2.ContractConfig{
+	destinationConfig := scribeConfig.ContractConfig{
 		Address:    u.LightInboxOnDestination.Address().String(),
 		StartBlock: 0,
 	}
@@ -81,7 +82,7 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 		},
 		Contracts: []scribeConfig2.ContractConfig{destinationConfig},
 	}
-	summitConfig := scribeConfig2.ContractConfig{
+	summitConfig := scribeConfig.ContractConfig{
 		Address:    u.InboxOnSummit.Address().String(),
 		StartBlock: 0,
 	}
@@ -97,8 +98,8 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 		},
 		Contracts: []scribeConfig2.ContractConfig{summitConfig},
 	}
-	scribeConfig := scribeConfig2.Config{
-		Chains: []scribeConfig2.ChainConfig{originChainConfig, destinationChainConfig, summitChainConfig},
+	scribeConfig := scribeConfig.Config{
+		Chains: []scribeConfig.ChainConfig{originChainConfig, destinationChainConfig, summitChainConfig},
 	}
 	clients := map[uint32][]backfill.ScribeBackend{
 		uint32(u.TestBackendOrigin.GetChainID()):      {originClient, originClient},
@@ -127,11 +128,11 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 	destination := uint32(u.TestBackendDestination.GetChainID())
 	summit := uint32(u.TestBackendSummit.GetChainID())
 
-	excCfg := executorCfg.Config{
+	excCfg := execConfig.Config{
 		SummitChainID: summit,
 		SummitAddress: u.SummitContract.Address().String(),
 		InboxAddress:  u.InboxOnSummit.Address().String(),
-		Chains: []executorCfg.ChainConfig{
+		Chains: []execConfig.ChainConfig{
 			{
 				ChainID:       chainID,
 				OriginAddress: u.OriginContract.Address().String(),
@@ -146,8 +147,8 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 			},
 		},
 		BaseOmnirpcURL: u.TestBackendOrigin.RPCAddress(),
-		UnbondedSigner: config2.SignerConfig{
-			Type: config2.FileType.String(),
+		UnbondedSigner: signerConfig.SignerConfig{
+			Type: signerConfig.FileType.String(),
 			File: filet.TmpFile(u.T(), "", u.ExecutorUnbondedWallet.PrivateKeyHex()).Name(),
 		},
 	}
@@ -182,12 +183,12 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 		},
 		DomainID:       uint32(0),
 		SummitDomainID: u.SummitDomainClient.Config().DomainID,
-		BondedSigner: config2.SignerConfig{
-			Type: config2.FileType.String(),
+		BondedSigner: signerConfig.SignerConfig{
+			Type: signerConfig.FileType.String(),
 			File: filet.TmpFile(u.T(), "", u.GuardBondedWallet.PrivateKeyHex()).Name(),
 		},
-		UnbondedSigner: config2.SignerConfig{
-			Type: config2.FileType.String(),
+		UnbondedSigner: signerConfig.SignerConfig{
+			Type: signerConfig.FileType.String(),
 			File: filet.TmpFile(u.T(), "", u.GuardUnbondedWallet.PrivateKeyHex()).Name(),
 		},
 		RefreshIntervalSeconds: 5,
@@ -200,12 +201,12 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 		},
 		DomainID:       u.DestinationDomainClient.Config().DomainID,
 		SummitDomainID: u.SummitDomainClient.Config().DomainID,
-		BondedSigner: config2.SignerConfig{
-			Type: config2.FileType.String(),
+		BondedSigner: signerConfig.SignerConfig{
+			Type: signerConfig.FileType.String(),
 			File: filet.TmpFile(u.T(), "", u.NotaryBondedWallet.PrivateKeyHex()).Name(),
 		},
-		UnbondedSigner: config2.SignerConfig{
-			Type: config2.FileType.String(),
+		UnbondedSigner: signerConfig.SignerConfig{
+			Type: signerConfig.FileType.String(),
 			File: filet.TmpFile(u.T(), "", u.NotaryUnbondedWallet.PrivateKeyHex()).Name(),
 		},
 		RefreshIntervalSeconds: 5,
@@ -229,7 +230,12 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 
 	Equal(u.T(), encodedNotaryTestConfig, decodedAgentConfigBackToEncodedBytes)
 
-	guard, err := guard.NewGuard(u.GetTestContext(), guardTestConfig, u.GuardMetrics)
+	rpcURLs := map[uint32]string{
+		chainID:     u.TestBackendOrigin.RPCAddress(),
+		destination: u.TestBackendDestination.RPCAddress(),
+		summit:      u.TestBackendSummit.RPCAddress(),
+	}
+	guard, err := guard.NewGuardInjectedBackend(u.GetTestContext(), guardTestConfig, u.GuardMetrics, rpcURLs)
 	Nil(u.T(), err)
 
 	tips := types.NewTips(big.NewInt(0), big.NewInt(0), big.NewInt(0), big.NewInt(0))
@@ -307,7 +313,7 @@ func (u *AgentsIntegrationSuite) TestAgentsE2E() {
 		return state.Nonce() >= uint32(1)
 	})
 
-	notary, err := notary.NewNotary(u.GetTestContext(), notaryTestConfig, u.NotaryMetrics)
+	notary, err := notary.NewNotaryInjectedBackend(u.GetTestContext(), notaryTestConfig, u.NotaryMetrics, rpcURLs)
 	Nil(u.T(), err)
 
 	go func() {