+++ agents/agents/guard/cmd/commands.go
@@ -42,12 +42,6 @@ var metricsPortFlag = &cli.UintFlag{
 	Value: 0,
 }
 
-var ignoreInitErrorsFlag = &cli.BoolFlag{
-	Name:  "ignore-init-errors",
-	Usage: "--ignore-init-errors",
-	Value: false,
-}
-
 var configFlag = &cli.StringFlag{
 	Name:      "config",
 	Usage:     "--config /Users/synapsecns/guard_config.yaml",
@@ -59,7 +53,7 @@ var configFlag = &cli.StringFlag{
 var GuardRunCommand = &cli.Command{
 	Name:        "guard-run",
 	Description: "runs the guard service",
-	Flags:       []cli.Flag{configFlag, metricsPortFlag, ignoreInitErrorsFlag},
+	Flags:       []cli.Flag{configFlag, metricsPortFlag},
 	Action: func(c *cli.Context) error {
 		handler, err := metrics.NewFromEnv(c.Context, metadata.BuildInfo())
 		if err != nil {
@@ -79,7 +73,7 @@ var GuardRunCommand = &cli.Command{
 			g, _ := errgroup.WithContext(c.Context)
 
 			guard, err := guard.NewGuard(c.Context, guardConfig, handler)
-			if err != nil && !c.Bool(ignoreInitErrorsFlag.Name) {
+			if err != nil {
 				return fmt.Errorf("failed to create guard: %w", err)
 			}