+++ agents/agents/notary/notary.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/synapsecns/sanguine/core/metrics"
 
-	config2 "github.com/synapsecns/sanguine/ethergo/signer/config"
+	signerConfig "github.com/synapsecns/sanguine/ethergo/signer/config"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
 	"time"
@@ -47,19 +47,21 @@ func NewNotary(ctx context.Context, cfg config.AgentConfig, handler metrics.Hand
 	}
 	notary.domains = []domains.DomainClient{}
 
-	notary.bondedSigner, err = config2.SignerFromConfig(ctx, cfg.BondedSigner)
+	notary.bondedSigner, err = signerConfig.SignerFromConfig(ctx, cfg.BondedSigner)
 	if err != nil {
 		return Notary{}, fmt.Errorf("error with bondedSigner, could not create notary: %w", err)
 	}
 
-	notary.unbondedSigner, err = config2.SignerFromConfig(ctx, cfg.UnbondedSigner)
+	notary.unbondedSigner, err = signerConfig.SignerFromConfig(ctx, cfg.UnbondedSigner)
 	if err != nil {
 		return Notary{}, fmt.Errorf("error with unbondedSigner, could not create notary: %w", err)
 	}
 
 	for domainName, domain := range cfg.Domains {
 		var domainClient domains.DomainClient
-		domainClient, err = evm.NewEVM(ctx, domainName, domain)
+
+		chainRPCURL := fmt.Sprintf("%s/confirmations/1/rpc/%d", cfg.BaseOmnirpcURL, domain.DomainID)
+		domainClient, err = evm.NewEVM(ctx, domainName, domain, chainRPCURL)
 		if err != nil {
 			return Notary{}, fmt.Errorf("failing to create evm for domain, could not create notary for: %w", err)
 		}
@@ -93,6 +95,9 @@ func (n *Notary) loadSummitMyLatestStates(parentCtx context.Context) {
 		))
 
 		originID := domain.Config().DomainID
+		if n.destinationDomain.Config().DomainID == originID {
+			continue
+		}
 		myLatestState, err := n.summitDomain.Summit().GetLatestAgentState(ctx, originID, n.bondedSigner)
 		if err != nil {
 			myLatestState = nil
@@ -348,6 +353,9 @@ func (n *Notary) getLatestSnapshot(parentCtx context.Context) (types.Snapshot, m
 	snapshotStates := make([]types.State, 0, len(statesToSubmit))
 	for _, state := range statesToSubmit {
 		if state.Nonce() == 0 {
+			continue
+		}
+		if state.Origin() == n.destinationDomain.Config().DomainID {
 			continue
 		}
 		snapshotStates = append(snapshotStates, state)