+++ core/metrics/handlertype_string.go
@@ -9,14 +9,14 @@ func _() {
 	// Re-run the stringer command to generate them again.
 	var x [1]struct{}
 	_ = x[DataDog-1]
-	_ = x[NewRelic-2]
+	_ = x[OTLP-2]
 	_ = x[Jaeger-3]
 	_ = x[Null-4]
 }
 
-const _HandlerType_name = "DatadogNewRelicJaegerNull"
+const _HandlerType_name = "DatadogOTLPJaegerNull"
 
-var _HandlerType_index = [...]uint8{0, 7, 15, 21, 25}
+var _HandlerType_index = [...]uint8{0, 7, 11, 17, 21}
 
 func (i HandlerType) String() string {
 	i -= 1