+++ ethergo/submitter/db/service.go
@@ -31,6 +31,10 @@ type Service interface {
 	DBTransaction(ctx context.Context, f TransactionFunc) error
 	// GetAllTXAttemptByStatus gets all txs for a given address and chain id with a given status.
 	GetAllTXAttemptByStatus(ctx context.Context, fromAddress common.Address, chainID *big.Int, matchStatuses ...Status) (txs []TX, err error)
+	// GetNonceStatus returns the nonce status for a given nonce by aggregating all attempts and finding the highest status.
+	GetNonceStatus(ctx context.Context, fromAddress common.Address, chainID *big.Int, nonce uint64) (status Status, err error)
+	// GetNonceAttemptsByStatus gets all txs for a given address and chain id with a given status and nonce.
+	GetNonceAttemptsByStatus(ctx context.Context, fromAddress common.Address, chainID *big.Int, nonce uint64, matchStatuses ...Status) (txs []TX, err error)
 }
 
 // TransactionFunc is a function that can be passed to DBTransaction.
@@ -43,6 +47,9 @@ type Status uint8
 
 // Important: do not modify the order of these constants.
 // if one needs to be removed, replace it with a no-op status.
+// additionally, due to the GetMaxNoncestatus function, statuses are currently assumed to be in order.
+// if you need to modify this functionality, please update that function. to reflect that the highest status
+// isno longer the expected end status.
 const (
 	// Pending is the status of a tx that has not been processed yet.
 	Pending Status = iota + 1 // Pending
@@ -62,6 +69,14 @@ const (
 
 var allStatusTypes = []Status{Pending, Stored, Submitted, FailedSubmit, ReplacedOrConfirmed, Replaced, Confirmed}
 
+// AllStatusTypes returns all status types.
+// it is exported for testing purposes
+//
+// These are guaranteed to be in order.
+func AllStatusTypes() []Status {
+	return allStatusTypes
+}
+
 // check to make sure all statuses are included in all status types.
 func _() {
 	for i := 0; i < len(_Status_index); i++ {
@@ -69,6 +84,9 @@ func _() {
 		status := Status(statusNum)
 		if status.String() == "" {
 			panic(fmt.Sprintf("invalid status: %d", status))
+		}
+		if status.String() != AllStatusTypes()[i].String() {
+			panic(fmt.Sprintf("status string and all status types do not match: %s, %s", status.String(), AllStatusTypes()[i]))
 		}
 	}
 }
@@ -107,5 +125,9 @@ func (s *Status) Value() (driver.Value, error) {
 
 var _ dbcommon.EnumInter = (*Status)(nil)
 
-// ErrNoNonceForChain is the error returned when there is no nonce for a given chain id.
-var ErrNoNonceForChain = errors.New("no nonce exists for this chain")
+var (
+	// ErrNoNonceForChain is the error returned when there is no nonce for a given chain id.
+	ErrNoNonceForChain = errors.New("no nonce exists for this chain")
+	// ErrNonceNotExist is the error returned when a nonce does not exist.
+	ErrNonceNotExist = errors.New("nonce does not exist")
+)