+++ packages/synapse-interface/constants/bridge.ts
@@ -114,6 +114,12 @@ export const BRIDGE_CONTRACTS = {
   [CHAINS.DOGE.id]: '0x9508BF380c1e6f751D97604732eF1Bae6673f299',
 }
 
+export const SYNAPSE_CCTP_CONTRACTS = {
+  [CHAINS.ETH.id]: '0xfB2Bfc368a7edfD51aa2cbEC513ad50edEa74E84',
+  [CHAINS.ARBITRUM.id]: '0xfB2Bfc368a7edfD51aa2cbEC513ad50edEa74E84',
+  [CHAINS.AVALANCHE.id]: '0xfB2Bfc368a7edfD51aa2cbEC513ad50edEa74E84',
+}
+
 export const ROLE_EVENTS = ['RoleGranted', 'RoleRevoked', 'RoleAdminChanged']
 
 export const INCLUDED_BRIDGE_EVENTS = [